ToneAlarm class refactoring to implement an interface for hardware specific methods and a single ToneAlarm class.
