+++ boards/airmind/mindpx-v2/default.cmake
@@ -45,6 +45,7 @@ px4_add_board(
 		tap_esc
 		telemetry # all available telemetry drivers
 		test_ppm
+		tone_alarm
 		uavcan
 
 	MODULES
