+++ boards/av/x-v1/default.cmake
@@ -50,6 +50,7 @@ px4_add_board(
 		tap_esc
 		telemetry # all available telemetry drivers
 		test_ppm
+		#tone_alarm
 		uavcan
 
 	MODULES
