+++ boards/gumstix/aerocore2/default.cmake
@@ -48,6 +48,7 @@ px4_add_board(
 		#tap_esc
 		#telemetry # all available telemetry drivers
 		#test_ppm
+		tone_alarm
 
 	MODULES
 		#attitude_estimator_q
