+++ boards/px4/fmu-v2/default.cmake
@@ -59,6 +59,7 @@ px4_add_board(
 		#tap_esc
 		#telemetry # all available telemetry drivers
 		#test_ppm
+		tone_alarm
 		#uavcan
 
 	MODULES
