+++ boards/px4/fmu-v3/stackcheck.cmake
@@ -58,6 +58,7 @@ px4_add_board(
 		tap_esc
 		telemetry # all available telemetry drivers
 		test_ppm
+		tone_alarm
 		#uavcan
 
 	MODULES
