+++ boards/px4/fmu-v4/stackcheck.cmake
@@ -43,6 +43,7 @@ px4_add_board(
 		tap_esc
 		telemetry # all available telemetry drivers
 		test_ppm
+		tone_alarm
 		#uavcan
 
 	MODULES
