+++ boards/px4/fmu-v4pro/default.cmake
@@ -57,6 +57,7 @@ px4_add_board(
 		tap_esc
 		telemetry # all available telemetry drivers
 		test_ppm
+		tone_alarm
 		uavcan
 
 	MODULES
