+++ boards/px4/sitl/rtps.cmake
@@ -17,7 +17,8 @@ px4_add_board(
 		#magnetometer # all available magnetometer drivers
 		pwm_out_sim
 		#telemetry # all available telemetry drivers
-		tone_alarm_sim
+		sim/tone_alarm
+		tone_alarm
 		#uavcan
 
 	MODULES
