+++ src/drivers/stm32/tone_alarm/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2015-2019 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,12 +31,6 @@
 #
 ############################################################################
 
-px4_add_module(
-	MODULE drivers__tone_alarm
-	MAIN tone_alarm
-	SRCS
-		tone_alarm.cpp
-	DEPENDS
-		circuit_breaker
-		tunes
-	)
+px4_add_library(tone_alarm_interface
+	ToneAlarmInterface.cpp
+)
