+++ src/systemcmds/tests/test_hrt.cpp
@@ -124,10 +124,10 @@ int test_tone(int argc, char *argv[])
 	int fd, result;
 	unsigned long tone;
 
-	fd = px4_open(TONEALARM0_DEVICE_PATH, O_WRONLY);
+	fd = px4_open(TONE_ALARM0_DEVICE_PATH, O_WRONLY);
 
 	if (fd < 0) {
-		printf("failed opening " TONEALARM0_DEVICE_PATH "\n");
+		printf("failed opening " TONE_ALARM0_DEVICE_PATH "\n");
 		goto out;
 	}
